env: split up Windows and Unix environment variable handling
